[Bugfix] KserveV2 Envelope May Increase Response Data Dimension
In kserve v2 envelope, if a request data is not an array (which
means it is not a batch) and the predict result is a string or an
integer, the result will be formated into an array of size 1, e.g.
[str] or [int].

In a single model, it simply creates an inconsistence between the
request and the response data dimension.

Within a workflow, it will break the chain.
For example:
    If I create a workflow of a dummy node simply return the input
    data, and a model node.
    If I send a request with data "abc" directly to the model's api,
    the model will receive [{"body": "abc"}]
    If I send the same request to the workflow's api, after the first
    dummy node, the data changes from "abc" to ["abc"]. The model will
    receive [{"body": ["abc"]}].