Updating index.md to fix 8 broken links (#2329)
* Update index.md

Update to fix a broken link in index.md where the trailing .md is cut off from the management_api.md. Added an anchor link to force the .md to show up.

* Update to index.md

Update to index.md to fix several links ending in .md that sphinx is breaking. Added anchor links to each link and a corresponding anchor in the affected doc. Tested locally and seems to be working.

* Update inference_api.md

* Updated typos

Fixed typos and updated wordslist.txt

* Update wordlist.txt

---------

Co-authored-by: sekyonda <7411+sekyonda@users.noreply.ghe.oculus-rep.com>
Co-authored-by: lxning <23464292+lxning@users.noreply.github.com>