Fixing FAQs doc per issue #2204 (#2351)
* Update index.md

Update to fix a broken link in index.md where the trailing .md is cut off from the management_api.md. Added an anchor link to force the .md to show up.

* Update to index.md

Update to index.md to fix several links ending in .md that sphinx is breaking. Added anchor links to each link and a corresponding anchor in the affected doc. Tested locally and seems to be working.

* Update inference_api.md

* Updated typos

Fixed typos and updated wordslist.txt

* Update wordlist.txt

* FAQs Updates

Updated a couple of broken links on the FAQ site per issue #2204

* updates to resolve links

* Update some links in index

Updated some links in index.md to go to the pytorch html page instead of github. This is a nicer fix for the .md sphinx issue

---------

Co-authored-by: sekyonda <7411+sekyonda@users.noreply.ghe.oculus-rep.com>
Co-authored-by: lxning <23464292+lxning@users.noreply.github.com>
Co-authored-by: Ankith Gunapal <agunapal@ischool.Berkeley.edu>