fix(kserve): ensure there's a default protocol configured
The current implementation retrieves the protocol to use from the
"PROTOCOL_VERSION" environment variable however there's no default value
which will trigger an error when served through Kserve as the base class
does a protocol check in the predict method that will fail with None.

The default protocol uses the same value as the base class.