Run torchserve in foreground
This removes the need to run tail to prevent docker exit, and ensures
that signals will be passed by shell to torchserve, allowing it to
terminate gracefully.

This change moves the tail call to be used only if not running in serve
(default) mode, allowing docker exit, so that the container can be
restarted by the orchestrator, instead of leaving a container that no
longer runs a torchserve instance (this improves UX with Kubernetes).